requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
